gbsystem corrigé

+++ Application/prise_service_old/lib/GBSystem_X_Developpement/widgets/HOME_QUICK_ACCESS_WIDGETS/GBSystem_agence_widget.dart
@@ -87,7 +87,7 @@ class _AgenceQuickAccessWidgetState extends State<AgenceQuickAccessWidget> {
                             duration: const Duration(milliseconds: 400),
                             curve: Curves.easeInOut);
                       },
-                      position: currentIndex.round().toInt(),
+                      position: currentIndex.round().toDouble(),
                     );
                   }),
                 ),
