+++ Application/prise_service_old/lib/GBSystem_X_Developpement/widgets/HOME_QUICK_ACCESS_WIDGETS/GBSystem_evaluation_non_terminer_widget.dart
@@ -91,7 +91,7 @@ class _EvaluationNonTerminerWidgetState
                             duration: const Duration(milliseconds: 400),
                             curve: Curves.easeInOut);
                       },
-                      position: currentIndex.round().toInt(),
+                      position: currentIndex.round().toDouble(),
                     );
                   }),
                 ),
